chore(ddd): reorganize directory structure for better clarity
