+++ configs/typescript/tsconfig.app.json
@@ -25,8 +25,6 @@
       "@src/*": ["../../src/*"],
       "@app/*": ["../../src/app/*"],
       "@configs/*": ["../../configs/*"],
-      "@server/*": ["../../src/server/*"],
-      "@modules/*": ["../../src/modules/*"],
       "@shared/*": ["../../src/shared/*"],
       "@unit-tests/*": ["../../tests/unit/*"],
       "@factories/*": ["../../tests/unit/utils/factories/*"],
