+++ src/infrastructure/api/health/types/health.types.ts
@@ -1,0 +1,13 @@
+import { createZodDto } from "nestjs-zod";
+
+import { HEALTH_CHECK_RESULT_SCHEMA } from "@src/infrastructure/api/health/constants/health.constants";
+
+import type { z } from "zod";
+
+type AppHealthCheckResult = z.infer<typeof HEALTH_CHECK_RESULT_SCHEMA>;
+
+class GetAppHealthCheckResultResponseDto extends createZodDto(HEALTH_CHECK_RESULT_SCHEMA) {}
+
+export type { AppHealthCheckResult };
+
+export { GetAppHealthCheckResultResponseDto };
