+++ src/infrastructure/api/server/swagger/constants/swagger.enums.ts
