--- src/modules/database/database.module.ts
@@ -1,18 +1,0 @@
-import { Module } from "@nestjs/common";
-import { MongooseModule } from "@nestjs/mongoose";
-import { LoggerModule } from "nestjs-pino";
-
-import { DatabaseService } from "@modules/database/providers/database.service";
-
-@Module({
-  imports: [
-    LoggerModule,
-    MongooseModule.forRootAsync({
-      imports: [],
-      inject: [],
-      useClass: DatabaseService,
-    }),
-  ],
-})
-
-export class DatabaseModule {}
