+++ tests/acceptance/features/step-definitions/app/docs/app-docs.then-steps.ts
@@ -1,7 +1,7 @@
 import { Then } from "@cucumber/cucumber";
 import { expect } from "expect";
 
-import { SWAGGER_DOCUMENTATION_TITLE } from "@server/constants/swagger.constants";
+import { SWAGGER_DOCUMENTATION_TITLE } from "@src/infrastructure/api/server/swagger/constants/swagger.constants";
 
 import type { GoatItWorld } from "@acceptance-support/types/world.types";
