+++ tests/unit/utils/mocks/infrastructure/api/health/providers/services/health.service.mock.ts
@@ -1,0 +1,16 @@
+import { createFakeHealthCheckResult } from "@factories/infrastructure/api/health/health.factory";
+
+import type { HealthCheckResult } from "@nestjs/terminus";
+import type { Mock } from "vitest";
+
+type MockedHealthService = {
+  checkAppHealth: Mock<() => Promise<HealthCheckResult>>;
+};
+
+function createMockedHealthService(): MockedHealthService {
+  return {
+    checkAppHealth: vi.fn<() => Promise<HealthCheckResult>>().mockResolvedValue(createFakeHealthCheckResult()),
+  };
+}
+
+export { createMockedHealthService };
